Update BinaryTree.java
the original code is missing Math.floor() around the log of size+1. this results in a leafCount of 0 regardless of input. as a result the code will not perform the initial rotations as defined by the DSW algorithm which in turn means the tree will not be properly balanced. your project probably still functions but this algorithm does not work quite as intended.

this first call to compression creates the "deepest nodes" and without this call the code will go right into the size/2 decrementing while loop.. the tree will probably still function for search but it won't be optimized. you want the "deepest nodes" case handled because these are the nodes that would dangle at the bottom of an otherwise perfect tree 

huh? ok, here's an example: if you has 10 nodes (s=10) you first call to compression should do 3 rotations then 10-3 leaves you with (s=7). the loop will dived, 7/2 = (in)3.5 = 3 so compression will do another 3 rotations, 3/2 = (int)1.5 = 1 so compression will do one final rotation. but the code as stands would have a leafCount of 0 at the start and would jump straight into the 5 rotations as called by s=10/2, then 2, then 1. this results in a sub-optimally balanced tree. in some cases, say 9 nodes, you will have an unbalanced tree.